% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOND.R
\name{BOND}
\alias{BOND}
\title{Bond Pricing Tool For Any Component
Specify which part of the bond calculation should be made and input the other variables.}
\usage{
BOND(
  price = NA,
  yield = NA,
  coupon_rate = NA,
  face_value = 100,
  maturity = NA,
  solve_for = "price"
)
}
\arguments{
\item{price}{The price of the bond.}

\item{yield}{The current yield of the bond.}

\item{coupon_rate}{The annualized coupon rate.}

\item{maturity}{Years to maturity.}

\item{solve_for}{The variable the function should solve for.}
}
\value{
A component of bond pricing defined by the user inputs.
}
\description{
Bond Pricing Tool For Any Component
Specify which part of the bond calculation should be made and input the other variables.
}
\examples{
BOND(yield = 0.05, coupon_rate = 0.06, maturity = 3, solve_for = "price")
BOND(price = 95, coupon_rate = 0.06, maturity = 3, solve_for = "yield")
BOND(price = 95, yield = 0.05, maturity = 3, solve_for = "coupon_rate")
BOND(price = 95, yield = 0.05, coupon_rate = 0.06, maturity = 3, solve_for = "face_value")
}
