% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTSTRAP_S.R
\name{BOOTSTRAP_S}
\alias{BOOTSTRAP_S}
\title{Bootstrap Yield Curve from Swap Rates}
\usage{
BOOTSTRAP_S(data)
}
\arguments{
\item{data}{A data frame with swap rates and maturities.}
}
\value{
A tibble with maturities and zero rates.
}
\description{
This function bootstraps a zero-coupon yield curve from swap rates.
It assumes annualized par rates
}
\examples{
swap_data <- tibble(maturity = c(1, 2, 3, 4),
                    rate = c(0.03, 0.035, 0.04, 0.045)
BOOTSTRAP_S(data=swap_data)
}
