% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM.R
\name{GBM}
\alias{GBM}
\title{Geometric Brownian Motion Simulation}
\usage{
GBM(spot, drift, sigma, T2M, delta_T, nsims)
}
\arguments{
\item{spot}{Initial spot price}

\item{drift}{Drift rate}

\item{sigma}{Volatility}

\item{T2M}{Time to maturity}

\item{delta_T}{Time step}

\item{nsims}{Number of simulations}
}
\value{
A tibble with simulated paths
}
\description{
This function simulates a Geometric Brownian Motion.It should not be used in interest rate models.
Instead it can be used in:
Inflation-linked bonds
Commodity-linked bonds
Credit spreads in high-yield debt
Convertible bonds (stock modeling component)
}
\examples{
GBM(spot=10, drift=0, sigma=0.2, T2M=1, delta_T=1/12, nsims=2)
}
