% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.R
\name{HW}
\alias{HW}
\title{Hull-White One-Factor Model Interest Rate Simulation}
\usage{
HW(LT_mean, rate, delta_T, sigma, theta, T2M, nsims)
}
\arguments{
\item{LT_mean}{Long-term mean function (constant or function of time)}

\item{rate}{Initial interest rate}

\item{delta_T}{Time step}

\item{sigma}{Volatility of interest rates}

\item{theta}{Mean reversion speed}

\item{T2M}{Time to maturity (total time horizon)}

\item{nsims}{Number of simulation paths}
}
\value{
A tibble with simulated interest rate paths
}
\description{
This function simulates the Hull-White one-factor model, which allows for a time-dependent mean.
The model is widely used in fixed-income derivatives pricing (e.g., swaptions, bond options).
}
\examples{
HW(LT_mean = function(t) 0.05, rate = 0.03, delta_T = 1/12, sigma = 0.02, theta = 0.1, T2M = 1, nsims = 5)
}
