% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAYLOR_PL.R
\name{TAYLOR_PL}
\alias{TAYLOR_PL}
\title{Taylor Series Expansion Profit/Loss}
\usage{
TAYLOR_PL(delta, gamma, theta)
}
\arguments{
\item{delta}{Delta/Durration value (First-Order)}

\item{gamma}{Gamma/Convexity value (Second-Order)}

\item{theta}{Theta/Speed value (Third-Order)}
}
\value{
The Profit/Loss Given The Taylor Series Expansion
}
\description{
This function measures profit and loss given Delta (Duration), Gamma (Convexity), and Speed
}
\examples{
vasicek(LT_mean = 0.05, rate = 0.03, delta_T = 1/12, sigma = 0.02, theta = 0.1, T2M = 1, nsims = 5)
}
