% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREASURY_US.R
\name{TREASURY_US}
\alias{TREASURY_US}
\title{US Treasury Data}
\usage{
TREASURY_US()
}
\value{
A tibble with rate data
}
\description{
This fucntion pulls rate data directly from the US Federal Reserve (As early as Jan-1962)
}
\examples{
TREASURY_US()
}
