% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VASICEK.R
\name{vasicek}
\alias{vasicek}
\title{Vasicek Interest Rate Simulation}
\usage{
vasicek(LT_mean, rate, delta_T, sigma, theta, T2M, nsims)
}
\arguments{
\item{LT_mean}{Long-term mean interest rate (mu)}

\item{rate}{Initial interest rate}

\item{delta_T}{Time step}

\item{sigma}{Volatility of interest rates}

\item{theta}{Mean reversion speed}

\item{T2M}{Time to maturity (total time horizon)}

\item{nsims}{Number of simulation paths}
}
\value{
A tibble with simulated interest rate paths
}
\description{
This function simulates the Vasicek model, which is commonly used for interest rate modeling.
}
\examples{
vasicek(LT_mean = 0.05, rate = 0.03, delta_T = 1/12, sigma = 0.02, theta = 0.1, T2M = 1, nsims = 5)
}
